pyproject.toml
[tool.poetry]
name = "prompt-optimizer"
version = "0.1.0"
description = ""
authors = ["Azfar Khoja <akhoja@coveo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

python-dotenv = "1.2.1"

langchain-openai = "0.3.35"
langchain-aws = "0.2.35"

pandas = "2.3.3"

jupyterlab = "4.5.0"
ipywidgets = "8.1.7"
matplotlib = "3.8.0"

pytest = "8.4.1"
mypy = "1.17.1"
ruff = "0.12.9"

[tool.ruff]
line-length = 120

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
